# docs/swagger.yaml
basePath: /api
definitions:
  github_com_amalgamated-tools_biblioteka_internal_db.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
  internal_handlers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handlers.seriesDTO:
    properties:
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
  internal_handlers.seriesRequest:
    properties:
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal library management API
  title: Biblioteka API
  version: "1.0"
paths:
...
  /series:
    get:
      description: Returns all series
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.seriesDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List series
      tags:
      - Series
    post:
      consumes:
      - application/json
      description: Create a new series
      parameters:
      - description: Series data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.seriesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.seriesDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a series
      tags:
      - Series
  /series/{id}:
    delete:
      description: Delete a series by ID
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a series
      tags:
      - Series
    get:
      description: Returns a single series by ID
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.seriesDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a series
      tags:
      - Series
    put:
      consumes:
      - application/json
      description: Update an existing series
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: string
      - description: Series data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.seriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.seriesDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a series
      tags:
      - Series
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"